fix: handle symlinked crash directories
